Infrastructure/GoogleApi/requirements.txt
google-cloud-bigquery
google-cloud-core
google-cloud-storage
